+++ provider/resource_custom_form.go
@@ -1,0 +1,179 @@
+package provider
+
+import (
+	"context"
+	"fmt"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/rootlyhq/terraform-provider-rootly/client"
+	"github.com/rootlyhq/terraform-provider-rootly/tools"
+)
+
+func resourceCustomForm() *schema.Resource {
+	return &schema.Resource{
+		CreateContext: resourceCustomFormCreate,
+		ReadContext:   resourceCustomFormRead,
+		UpdateContext: resourceCustomFormUpdate,
+		DeleteContext: resourceCustomFormDelete,
+		Importer: &schema.ResourceImporter{
+			StateContext: schema.ImportStatePassthroughContext,
+		},
+		Schema: map[string]*schema.Schema{
+
+			"name": &schema.Schema{
+				Type:        schema.TypeString,
+				Computed:    false,
+				Required:    true,
+				Optional:    false,
+				ForceNew:    false,
+				Description: "The name of the custom form.",
+			},
+
+			"slug": &schema.Schema{
+				Type:        schema.TypeString,
+				Computed:    true,
+				Required:    false,
+				Optional:    true,
+				ForceNew:    false,
+				Description: "The custom form slug. Add this to form_field.shown or form_field.required to associate form fields with custom forms.",
+			},
+
+			"description": &schema.Schema{
+				Type:        schema.TypeString,
+				Computed:    true,
+				Required:    false,
+				Optional:    true,
+				ForceNew:    false,
+				Description: "",
+			},
+
+			"enabled": &schema.Schema{
+				Type:     schema.TypeBool,
+				Default:  true,
+				Optional: true,
+			},
+
+			"command": &schema.Schema{
+				Type:        schema.TypeString,
+				Computed:    false,
+				Required:    true,
+				Optional:    false,
+				ForceNew:    false,
+				Description: "The Slack command used to trigger this form.",
+			},
+		},
+	}
+}
+
+func resourceCustomFormCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	c := meta.(*client.Client)
+
+	tflog.Trace(ctx, fmt.Sprintf("Creating CustomForm"))
+
+	s := &client.CustomForm{}
+
+	if value, ok := d.GetOkExists("name"); ok {
+		s.Name = value.(string)
+	}
+	if value, ok := d.GetOkExists("slug"); ok {
+		s.Slug = value.(string)
+	}
+	if value, ok := d.GetOkExists("description"); ok {
+		s.Description = value.(string)
+	}
+	if value, ok := d.GetOkExists("enabled"); ok {
+		s.Enabled = tools.Bool(value.(bool))
+	}
+	if value, ok := d.GetOkExists("command"); ok {
+		s.Command = value.(string)
+	}
+
+	res, err := c.CreateCustomForm(s)
+	if err != nil {
+		return diag.Errorf("Error creating custom_form: %s", err.Error())
+	}
+
+	d.SetId(res.ID)
+	tflog.Trace(ctx, fmt.Sprintf("created a custom_form resource: %s", d.Id()))
+
+	return resourceCustomFormRead(ctx, d, meta)
+}
+
+func resourceCustomFormRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	c := meta.(*client.Client)
+	tflog.Trace(ctx, fmt.Sprintf("Reading CustomForm: %s", d.Id()))
+
+	item, err := c.GetCustomForm(d.Id())
+	if err != nil {
+		// In the case of a NotFoundError, it means the resource may have been removed upstream
+		// We just remove it from the state.
+		if _, ok := err.(client.NotFoundError); ok && !d.IsNewResource() {
+			tflog.Warn(ctx, fmt.Sprintf("CustomForm (%s) not found, removing from state", d.Id()))
+			d.SetId("")
+			return nil
+		}
+
+		return diag.Errorf("Error reading custom_form: %s", d.Id())
+	}
+
+	d.Set("name", item.Name)
+	d.Set("slug", item.Slug)
+	d.Set("description", item.Description)
+	d.Set("enabled", item.Enabled)
+	d.Set("command", item.Command)
+
+	return nil
+}
+
+func resourceCustomFormUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	c := meta.(*client.Client)
+	tflog.Trace(ctx, fmt.Sprintf("Updating CustomForm: %s", d.Id()))
+
+	s := &client.CustomForm{}
+
+	if d.HasChange("name") {
+		s.Name = d.Get("name").(string)
+	}
+	if d.HasChange("slug") {
+		s.Slug = d.Get("slug").(string)
+	}
+	if d.HasChange("description") {
+		s.Description = d.Get("description").(string)
+	}
+	if d.HasChange("enabled") {
+		s.Enabled = tools.Bool(d.Get("enabled").(bool))
+	}
+	if d.HasChange("command") {
+		s.Command = d.Get("command").(string)
+	}
+
+	_, err := c.UpdateCustomForm(d.Id(), s)
+	if err != nil {
+		return diag.Errorf("Error updating custom_form: %s", err.Error())
+	}
+
+	return resourceCustomFormRead(ctx, d, meta)
+}
+
+func resourceCustomFormDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	c := meta.(*client.Client)
+	tflog.Trace(ctx, fmt.Sprintf("Deleting CustomForm: %s", d.Id()))
+
+	err := c.DeleteCustomForm(d.Id())
+	if err != nil {
+		// In the case of a NotFoundError, it means the resource may have been removed upstream.
+		// We just remove it from the state.
+		if _, ok := err.(client.NotFoundError); ok && !d.IsNewResource() {
+			tflog.Warn(ctx, fmt.Sprintf("CustomForm (%s) not found, removing from state", d.Id()))
+			d.SetId("")
+			return nil
+		}
+		return diag.Errorf("Error deleting custom_form: %s", err.Error())
+	}
+
+	d.SetId("")
+
+	return nil
+}