+++ provider/resource_custom_form_test.go
@@ -1,0 +1,28 @@
+package provider
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+)
+
+func TestAccResourceCustomForm(t *testing.T) {
+	resource.UnitTest(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		ProviderFactories: providerFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccResourceCustomForm,
+			},
+		},
+	})
+}
+
+const testAccResourceCustomForm = `
+resource "rootly_custom_form" "test" {
+	name = "test"
+command = "test"
+}
+`