+++ provider/resource_post_mortem_template.go
@@ -40,11 +40,11 @@ func resourcePostmortemTemplate() *schema.Resource {
 			},
 
 			"content": &schema.Schema{
-				Type:        schema.TypeString,
-				Computed:    true,
-				Required:    false,
-				Optional:    true,
-				ForceNew:    true,
+				Type:     schema.TypeString,
+				Computed: true,
+				Required: false,
+				Optional: true,
+				ForceNew: true,
 				DiffSuppressFunc: func(key, oldValue string, newValue string, d *schema.ResourceData) bool {
 					if oldValue == "" {
 						return false